Add directory structure on README
- Move the transformations on useCountry hook
