+++ src/contexts/domain.tsx
@@ -3,26 +3,13 @@ import { paths } from '#generated/types';
 
 export type CacheKey = 'country' | 'global-enums'
 
-type GetGlobalEnums = paths['/api/v2/global-enums/']['get'];
-export type GlobalEnums = GetGlobalEnums['responses']['200']['content']['application/json'];
-
-type CountriesResponse = paths['/api/v2/country/']['get']['responses']['200']['content']['application/json'];
-export type Country = Omit<
-    NonNullable<CountriesResponse['results']>[number],
-    'id' | 'iso3' | 'iso3' | 'name' | 'is_deprecated' | 'independent'
-> & {
-    id: number;
-    iso: string;
-    iso3: string;
-    name: string;
-    is_deprecated: false | undefined;
-    independent: true;
-};
+export type GlobalEnums = paths['/api/v2/global-enums/']['get']['responses']['200']['content']['application/json'];
+export type Countries = paths['/api/v2/country/']['get']['responses']['200']['content']['application/json'];
 
 export interface Domain {
     register: (name: CacheKey) => void;
 
-    countries?: Country[];
+    countries?: Countries;
     countriesPending?: boolean;
 
     globalEnums?: GlobalEnums;
