+++ src/hooks/domain/useCountry.ts
@@ -1,13 +1,27 @@
 import {
     isDefined,
     isNotDefined,
+    isFalsyString,
 } from '@togglecorp/fujs';
 import {
+    useMemo,
     useContext,
     useEffect,
 } from 'react';
 
-import DomainContext, { Country } from '#contexts/domain';
+import DomainContext, { type Countries } from '#contexts/domain';
+
+type Country = Omit<
+    NonNullable<Countries['results']>[number],
+    'id' | 'iso3' | 'iso3' | 'name' | 'is_deprecated' | 'independent'
+> & {
+    id: number;
+    iso: string;
+    iso3: string;
+    name: string;
+    is_deprecated: false | undefined;
+    independent: true;
+};
 
 type ListProps = {
     region?: number;
@@ -33,13 +47,40 @@ function useCountry(props: PropsForIso3): Country | undefined
 function useCountry(
     props?: ListProps | PropsForId | PropsForIso3,
 ): (Country | undefined | Array<Country>) {
-    const { countries, register } = useContext(DomainContext);
+    const { countries: countriesUnsafe, register } = useContext(DomainContext);
 
     useEffect(
         () => {
             register('country');
         },
         [register],
+    );
+
+    const countries = useMemo(
+        () => (
+            countriesUnsafe?.results?.map((country) => {
+                if (isNotDefined(country.id)
+                    || isNotDefined(country.iso)
+                    || isNotDefined(country.iso3)
+                    || isFalsyString(country.name)
+                    || !!country.is_deprecated
+                    || !country.independent
+                ) {
+                    return undefined;
+                }
+
+                return {
+                    ...country,
+                    id: country.id,
+                    iso: country.iso,
+                    iso3: country.iso3,
+                    name: country.name,
+                    is_deprecated: country.is_deprecated,
+                    independent: country.independent,
+                };
+            }).filter(isDefined)
+        ),
+        [countriesUnsafe],
     );
 
     if (isNotDefined(props)) {
