+++ src/views/GlobalThreeW/Filters/index.tsx
@@ -18,7 +18,6 @@ import i18n from './i18n.json';
 import styles from './styles.module.css';
 
 type CountryListItem = NonNullable<GoApiResponse<'/api/v2/country/'>['results']>[number];
-// type GlobalEnumsResponse = GoApiResponse<'/api/v2/global-enums/'>;
 
 export interface FilterValue {
     reporting_ns: number[];
